requirements.txt
selenium
yaml
argparse
python-dateutil
requests
pathlib
numpy
shutil